Adding custom promises like dialog service and modifying toast service

+++ src/app/users/user-roles/user-role.service.ts
@@ -5,6 +5,7 @@ import { Observable } from 'rxjs';
 import { UserRole } from './redux/roles.state';
 import { selectAllUserRoles } from './redux/roles.selectors';
 import { ToastService } from '../shared/toast/toast.service';
+import { DialogService } from '../shared/dialog/dialog.service';
 
 @Injectable({
   providedIn: 'root',
@@ -12,6 +13,7 @@ import { ToastService } from '../shared/toast/toast.service';
 export class UserRoleService {
   private store = inject(Store);
   private toastService = inject(ToastService);
+  private dialogService = inject(DialogService);
   constructor() {}
 
   selectAllUserRoles(): Observable<UserRole[]> {
@@ -19,16 +21,24 @@ export class UserRoleService {
   }
   async addUserRole(newRole: any): Promise<void> {
     this.store!.dispatch(addUserRole({ role: newRole }));
-    this.toastService.displayMessage(
+    await this.toastService.displayMessage(
       `Role ${newRole.roleName} added successfully!`,
       'create'
     );
   }
   async deleteRole(roleName: string): Promise<void> {
-    this.store!.dispatch(deleteRole({ roleName }));
-    this.toastService.displayMessage(
-      `Role ${roleName} deleted successfully!`,
-      'delete'
+    const confirmed = await this.dialogService.confirm(
+      `Are you sure you want to delete the role ${roleName}?`
     );
+
+    if (confirmed) {
+      this.store!.dispatch(deleteRole({ roleName }));
+      await this.toastService.displayMessage(
+        `Role ${roleName} deleted successfully!`,
+        'delete'
+      );
+    } else {
+      console.log('Role deletion cancelled');
+    }
   }
 }
