+++ src/app/users/users.service.ts
@@ -6,12 +6,13 @@ import { mapResultToUser } from './utils/user.utils';
 import { selectAllUsers, selectUserById } from './redux/user.selectors';
 import { Observable } from 'rxjs';
 import { ToastService } from './shared/toast/toast.service';
+import { DialogService } from './shared/dialog/dialog.service';
 
 @Injectable({ providedIn: 'root' })
 export class UsersService {
   private store = inject(Store);
   private toastService = inject(ToastService);
-
+  private dialogService = inject(DialogService);
   getAllUsers(): Observable<User[]> {
     return this.store.select(selectAllUsers);
   }
@@ -19,7 +20,7 @@ export class UsersService {
   async addUser(data: any): Promise<void> {
     const newUser = mapResultToUser(data);
     this.store.dispatch(addUser({ user: newUser }));
-    this.toastService.displayMessage(
+    await this.toastService.displayMessage(
       `User ${newUser.name} added successfully!`,
       'create'
     );
@@ -28,18 +29,25 @@ export class UsersService {
   async updateUser(data: any): Promise<void> {
     const updatedUser = mapResultToUser(data, data.id);
     this.store.dispatch(updateUser({ user: updatedUser }));
-    this.toastService.displayMessage(
+    await this.toastService.displayMessage(
       `User ${updatedUser.name} updated successfully!`,
       'update'
     );
   }
 
   async deleteUser(userId: string): Promise<void> {
-    this.store.dispatch(deleteUser({ deleteUserId: userId }));
-    this.toastService.displayMessage(
-      `User with ID ${userId} deleted successfully!`,
-      'delete'
+    const confirmed = await this.dialogService.confirm(
+      `Are you sure you want to delete user with id:${userId} ?`
     );
+    if (confirmed) {
+      this.store.dispatch(deleteUser({ deleteUserId: userId }));
+      await this.toastService.displayMessage(
+        `User with ID ${userId} deleted successfully!`,
+        'delete'
+      );
+    } else {
+      console.log('User deletion cancelled');
+    }
   }
 
   showUserProfile(userId: string): Observable<User | undefined> {
